requirements.txt
absl-py==1.4.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.34
einops==0.6.0
fairscale==0.4.13
filelock==3.11.0
flash-attn==1.0.9
Flask==2.3.2
fonttools==4.39.3
future==0.18.3
gdown==4.7.1
google-auth==2.17.2
google-auth-oauthlib==1.0.0
grpcio==1.53.0
h5py==3.8.0
huggingface-hub==0.13.4
idna==3.4
imageio==2.27.0
importlib-metadata==6.3.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
lit==16.0.0
lmdb==1.4.1
lpips==0.1.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
MinkowskiEngine==0.5.4
mpmath==1.3.0
natsort==8.3.1
natten @ file:///home/lab/code1/natten-0.14.6%2Btorch200cu117-cp38-cp38-linux_x86_64.whl
networkx==3.1
ninja==1.11.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.0
Pillow==9.5.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
soupsieve==2.4
sympy==1.11.1
tb-nightly==2.13.0a20230409
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.6.13
torch==2.0.0
torchaudio==2.0.1
torchprofile==0.0.4
torchsummary==1.5.1
torchvision==0.15.1
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.6
yacs==0.1.8
yapf==0.32.0
zipp==3.15.0
